requirements.txt
asgiref==3.8.1
django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.6
sqlparse==0.5.1
